move re-usable code to cmdline module
